Moving all banning methods into bans model
* Moving all ban related code into the new bans model.

* Updating old code where necessary. Moving from adminUsers to bans.
