+++ server/plugins/moderation_log/templates.js
@@ -177,7 +177,31 @@ module.exports = {
       .then(function(user) { data.username = user.username; });
     }
   },
-  'adminUsers.ban': {
+
+  // =========== Banning Routes ===========
+  'bans.addAddresses': {
+    genDisplayText: function(data) {
+      var addresses = [];
+      data.addresses.forEach(function(addrInfo) {
+        addresses.push(addrInfo.hostname.toString().replace(/%/g, '*') || addrInfo.ip);
+      });
+      return `banned the following addresses "${addresses.toString().replace(/,/g, ', ')}"`;
+    },
+    genDisplayUrl: function() { return `admin-management.banned-addresses`; },
+  },
+  'bans.editAddress': {
+    genDisplayText: function(data) {
+      return `edited banned address "${data.hostname.toString().replace(/%/g, '*') || data.ip}" to ${data.decay ? 'decay' : 'not decay'} with a weight of ${data.weight}`;
+    },
+    genDisplayUrl: function(data) { return `admin-management.banned-addresses({ search: '${data.hostname || data.ip}' })`; },
+  },
+  'bans.deleteAddress': {
+    genDisplayText: function(data) {
+      return `deleted banned address "${data.hostname.toString().replace(/%/g, '*') || data.ip}"`;
+    },
+    genDisplayUrl: function() { return `admin-management.banned-addresses`; },
+  },
+  'bans.ban': {
     genDisplayText: function(data) {
       var humanDate;
       if (data.expiration) {
@@ -196,7 +220,7 @@ module.exports = {
       .then(function(user) { data.username = user.username; });
     }
   },
-  'adminUsers.unban': {
+  'bans.unban': {
     genDisplayText: function(data) { return `unbanned user "${data.username}"`; },
     genDisplayUrl: function(data) { return `profile({ username: '${data.username}' })`; },
     dataQuery: function(data, request) {
@@ -204,7 +228,7 @@ module.exports = {
       .then(function(user) { data.username = user.username; });
     }
   },
-  'adminUsers.banFromBoards': {
+  'bans.banFromBoards': {
     genDisplayText: function(data) {
       return `banned user "${data.username}" from boards: "${data.boards}"`;
     },
@@ -223,7 +247,7 @@ module.exports = {
       });
     }
   },
-  'adminUsers.unbanFromBoards': {
+  'bans.unbanFromBoards': {
     genDisplayText: function(data) {
       return `unbanned user "${data.username}" from boards: "${data.boards}"`;
     },
